+++ apps/Original/priv/gettext/en/LC_MESSAGES/errors.po
