+++ apps/Original/priv/gettext/errors.pot
